ObjectStore cleanup (#2587)
