+++ object_store/src/azure/credential.rs
@@ -15,19 +15,24 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::client::oauth::ClientSecretOAuthProvider;
+use crate::client::retry::RetryExt;
+use crate::client::token::{TemporaryToken, TokenCache};
 use crate::util::hmac_sha256;
+use crate::RetryConfig;
 use chrono::Utc;
+use reqwest::header::ACCEPT;
 use reqwest::{
     header::{
         HeaderMap, HeaderName, HeaderValue, AUTHORIZATION, CONTENT_ENCODING,
         CONTENT_LANGUAGE, CONTENT_LENGTH, CONTENT_TYPE, DATE, IF_MATCH,
         IF_MODIFIED_SINCE, IF_NONE_MATCH, IF_UNMODIFIED_SINCE, RANGE,
     },
-    Method, RequestBuilder,
+    Client, Method, RequestBuilder,
 };
+use snafu::{ResultExt, Snafu};
 use std::borrow::Cow;
 use std::str;
+use std::time::{Duration, Instant};
 use url::Url;
 
 static AZURE_VERSION: HeaderValue = HeaderValue::from_static("2021-08-06");
@@ -38,6 +43,18 @@ pub(crate) static DELETE_SNAPSHOTS: HeaderName =
 pub(crate) static COPY_SOURCE: HeaderName = HeaderName::from_static("x-ms-copy-source");
 static CONTENT_MD5: HeaderName = HeaderName::from_static("content-md5");
 pub(crate) static RFC1123_FMT: &str = "%a, %d %h %Y %T GMT";
+const CONTENT_TYPE_JSON: &str = "application/json";
+
+#[derive(Debug, Snafu)]
+pub enum Error {
+    #[snafu(display("Error performing token request: {}", source))]
+    TokenRequest { source: crate::client::retry::Error },
+
+    #[snafu(display("Error getting token response body: {}", source))]
+    TokenResponseBody { source: reqwest::Error },
+}
+
+pub type Result<T, E = Error> = std::result::Result<T, E>;
 
 /// Provides credentials for use when signing requests
 #[derive(Debug)]
@@ -252,4 +269,82 @@ fn lexy_sort<'a>(
         .collect::<Vec<_>>();
     values.sort_unstable();
     values
+}
+
+#[derive(serde::Deserialize, Debug)]
+struct TokenResponse {
+    access_token: String,
+    expires_in: u64,
+}
+
+/// Encapsulates the logic to perform an OAuth token challenge
+#[derive(Debug)]
+pub struct ClientSecretOAuthProvider {
+    scope: String,
+    token_url: String,
+    client_id: String,
+    client_secret: String,
+    cache: TokenCache<String>,
+}
+
+impl ClientSecretOAuthProvider {
+    /// Create a new [`ClientSecretOAuthProvider`] for an azure backed store
+    pub fn new(
+        client_id: String,
+        client_secret: String,
+        tenant_id: String,
+        authority_host: Option<String>,
+    ) -> Self {
+        let authority_host = authority_host
+            .unwrap_or_else(|| authority_hosts::AZURE_PUBLIC_CLOUD.to_owned());
+
+        Self {
+            scope: "https://storage.azure.com/.default".to_owned(),
+            token_url: format!("{}/{}/oauth2/v2.0/token", authority_host, tenant_id),
+            client_id,
+            client_secret,
+            cache: TokenCache::default(),
+        }
+    }
+
+    /// Fetch a token
+    pub async fn fetch_token(
+        &self,
+        client: &Client,
+        retry: &RetryConfig,
+    ) -> Result<String> {
+        self.cache
+            .get_or_insert_with(|| self.fetch_token_inner(client, retry))
+            .await
+    }
+
+    /// Fetch a fresh token
+    async fn fetch_token_inner(
+        &self,
+        client: &Client,
+        retry: &RetryConfig,
+    ) -> Result<TemporaryToken<String>> {
+        let response: TokenResponse = client
+            .request(Method::POST, &self.token_url)
+            .header(ACCEPT, HeaderValue::from_static(CONTENT_TYPE_JSON))
+            .form(&[
+                ("client_id", self.client_id.as_str()),
+                ("client_secret", self.client_secret.as_str()),
+                ("scope", self.scope.as_str()),
+                ("grant_type", "client_credentials"),
+            ])
+            .send_retry(retry)
+            .await
+            .context(TokenRequestSnafu)?
+            .json()
+            .await
+            .context(TokenResponseBodySnafu)?;
+
+        let token = TemporaryToken {
+            token: response.access_token,
+            expiry: Instant::now() + Duration::from_secs(response.expires_in),
+        };
+
+        Ok(token)
+    }
 }
