+++ object_store/src/client/mod.rs
@@ -18,8 +18,6 @@
 //! Generic utilities reqwest based ObjectStore implementations
 
 pub mod backoff;
-#[cfg(any(feature = "gcp", feature = "azure"))]
-pub mod oauth;
 pub mod pagination;
 pub mod retry;
 pub mod token;
